requirements.txt
click
flask
markdown2
emoji
python-dotenv
uwsgi